refactor split project-handlers >500 lines
