+++ services/web/server/tests/unit/with_dbs/02/test_projects_crud.py
@@ -10,11 +10,7 @@ from math import ceil
 from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union
 
 import pytest
-from _helpers import (  # type: ignore
-    ExpectedResponse,
-    MockedStorageSubsystem,
-    standard_role_response,
-)
+from _helpers import ExpectedResponse, MockedStorageSubsystem, standard_role_response
 from aiohttp import web
 from aiohttp.test_utils import TestClient
 from aioresponses import aioresponses
@@ -22,7 +18,7 @@ from models_library.projects_state import ProjectState
 from pytest_simcore.helpers.utils_assert import assert_status
 from simcore_service_webserver._meta import api_version_prefix
 from simcore_service_webserver.db_models import UserRole
-from simcore_service_webserver.projects.projects_handlers import (
+from simcore_service_webserver.projects.projects_handlers_crud import (
     OVERRIDABLE_DOCUMENT_KEYS,
 )
 from simcore_service_webserver.utils import now_str, to_datetime
